requirements.txt
flask 
google-generativeai 
python-dotenv
PyPDF2
python-docx